couple of changes
first I fixed an issue where if you weren't signed in the app would crash. For future reference. Don't try and access the demoV object directly. Just use the app level variable.

Upload now has a progress bar. For whatever reason it is screwy in the sense that it will freeze if the function is put on a 1 second timer. idk why

I do not know if the upload is done in the best way either, so don't sue me